Move Quaternion to synthesis module
