+++ qiskit/quantum_info/operators/__init__.py
@@ -16,7 +16,6 @@ from .operator import Operator
 from .scalar_op import ScalarOp
 from .pauli import Pauli, pauli_group
 from .channel import Choi, SuperOp, Kraus, Stinespring, Chi, PTM
-from .quaternion import Quaternion
 from .measures import (process_fidelity,
                        average_gate_fidelity,
                        gate_error,
