+++ qiskit/quantum_info/synthesis/__init__.py
@@ -14,3 +14,4 @@
 
 from .two_qubit_decompose import TwoQubitBasisDecomposer, two_qubit_cnot_decompose
 from .one_qubit_decompose import OneQubitEulerDecomposer
+from .quaternion import Quaternion
