+++ qiskit/quantum_info/synthesis/quaternion.py
