+++ qiskit/transpiler/passes/optimization/optimize_1q_gates.py
@@ -24,7 +24,7 @@ from qiskit.circuit.library.standard_gates.u2 import U2Gate
 from qiskit.circuit.library.standard_gates.u3 import U3Gate
 from qiskit.circuit.gate import Gate
 from qiskit.transpiler.basepasses import TransformationPass
-from qiskit.quantum_info.operators import Quaternion
+from qiskit.quantum_info.synthesis import Quaternion
 
 _CHOP_THRESHOLD = 1e-15
